+++ packages/terra-abstract-modal/src/_ModalContent.jsx
@@ -45,10 +45,6 @@ const propTypes = {
     PropTypes.func,
   ]),
   /**
-   * If set to true, the modal will trap the focus and prevents any popup within the modal from gaining focus.
-   */
-  isFocused: PropTypes.bool,
-  /**
    * If set to true, the modal will be fullscreen on all breakpoint sizes.
    */
   isFullscreen: PropTypes.bool,
@@ -71,7 +67,6 @@ const defaultProps = {
   classNameOverlay: null,
   closeOnOutsideClick: true,
   fallbackFocus: undefined,
-  isFocused: true,
   isFullscreen: false,
   isScrollable: false,
   role: 'dialog',
@@ -80,6 +75,19 @@ const defaultProps = {
 
 /* eslint-disable react/prefer-stateless-function */
 class ModalContent extends React.Component {
+  constructor() {
+    super();
+    this.modalContentRef = React.createRef();
+  }
+
+  componentDidMount() {
+    if (this.props.fallbackFocus) {
+      this.props.fallbackFocus.focus();
+    } else {
+      this.modalContentRef.current.focus();
+    }
+  }
+
   render() {
     const {
       ariaLabel,
@@ -90,7 +98,6 @@ class ModalContent extends React.Component {
       onRequestClose,
       fallbackFocus,
       role,
-      isFocused,
       isFullscreen,
       isScrollable,
       zIndex,
@@ -113,22 +120,8 @@ class ModalContent extends React.Component {
     delete customProps.closePortal;
     const platformIsiOS = !!navigator.platform && /iPad|iPhone|iPod/.test(navigator.platform);
 
-    this.modalContentRef = React.createRef();
-
-    if (fallbackFocus) {
-      this.fallbackFocus = fallbackFocus;
-    } else {
-      this.fallbackFocus = () => (this.modalContentRef.current);
-    }
-
     return (
-      <FocusTrap
-        paused={!isFocused}
-        focusTrapOptions={{
-          escapeDeactivates: false,
-          fallbackFocus: this.fallbackFocus,
-        }}
-      >
+      <React.Fragment>
         <ModalOverlay
           onClick={closeOnOutsideClick ? onRequestClose : null}
           className={classNameOverlay}
@@ -150,7 +143,7 @@ class ModalContent extends React.Component {
           {children}
         </div>
         {/* eslint-enable jsx-a11y/no-noninteractive-tabindex */}
-      </FocusTrap>
+      </React.Fragment>
     );
   }
 }